add: cache support to HLS player
The HLS player now supports caching. The cache directory is obtained from the context, and a cache is created using the Least Recently Used (LRU) eviction policy. The cached data will be used to play the video if available. If the data is not cached, it will be downloaded from the network. The cache is released when the player is released.