update: adjust bottom sheet padding for fullscreen mode
This commit introduces a new parameter `navigationBarHeightWhenHiddenDp` to `BottomSheetConfig` and updates `CustomModalBottomSheet` to add bottom padding when in portrait fullscreen mode. This prevents the bottom sheet from being obscured by the navigation bar.

The `VideoPlayer` has been updated to pass the `isFullscreen` state to the bottom sheets for subtitles and playback speed.